build-deploy.bat: signing with Halmer certificate VirtualList.java: workaround to eliminate empty window title for J9 VirtualList.java: searching for J9 hangup after Motorola threads fix

+++ src/ui/VirtualList.java
@@ -210,6 +210,10 @@ public abstract class VirtualList
     public VirtualList() {
         width=getWidth();
         height=getHeight();
+        
+        if (Info.Version.getPlatformName().startsWith("Windows")) {
+            setTitle("Bombus CE");
+        }
         // rotator
         //rotator=new TimerTaskRotate(0, this);
 //#if !(MIDP1)
@@ -892,6 +896,12 @@ class TimerTaskRotate extends Thread{
     }
     
     public static void startRotate(int max, VirtualList list){
+        //Windows mobile J9 hanging test
+        if (Info.Version.getPlatformName().startsWith("Windows")) {
+            list.showBalloon=true;
+            list.offset=0;
+            return;
+        }
         if (instance==null) instance=new TimerTaskRotate();
         if (max<0) {
             instance.destroyTask(); return;
